refactor(server): extract routes.rs + add OpenAPI conformance test (A.11)
Sprint 2 Wave 8 item A.11: Extract route builders into routes.rs module.
Add structural conformance test that sends real HTTP requests for every
OpenAPI-declared operation through the Axum router, catching orphaned or
missing routes at compile time.
